backend/requirements_exact.txt
# AI Eyes Surveillance System - Exact Dependencies for Python 3.10.1
# This file ensures zero conflicts on new laptop installations
# Generated for maximum compatibility and stability

# ========================================
# CORE WEB FRAMEWORK (Exact Versions)
# ========================================
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
eventlet==0.33.3
python-socketio==5.9.0

# ========================================
# AI/ML LIBRARIES (Conflict-Free Versions)
# ========================================
# TensorFlow 2.13.0 - Last version compatible with MediaPipe
tensorflow==2.13.0
# MediaPipe 0.10.21 - Latest stable for Python 3.10
mediapipe==0.10.21
# PyTorch stable for Python 3.10
torch==2.0.1
torchvision==0.15.2
# YOLOv8 specific version
ultralytics==8.0.196

# ========================================
# CRITICAL CONFLICT RESOLUTION
# ========================================
# Protobuf version MUST be exactly 3.20.3 for TF+MediaPipe compatibility
protobuf==3.20.3
# gRPC compatibility
grpcio==1.59.0
# Prevent setuptools warnings
setuptools==68.0.0

# ========================================
# COMPUTER VISION (Matched Versions)
# ========================================
# OpenCV with exact matching versions to prevent DLL conflicts
opencv-python==4.8.1.78
opencv-contrib-python==4.8.1.78
# NumPy compatible with all above packages
numpy==1.24.3
# Image processing
Pillow==10.0.1
imutils==0.5.4

# ========================================
# DATABASE & STORAGE
# ========================================
pymongo==4.5.0
motor==3.3.1

# ========================================
# COMMUNICATION & APIs
# ========================================
requests==2.31.0
sendgrid==6.10.0
python-dotenv==1.0.0

# ========================================
# SCIENTIFIC COMPUTING
# ========================================
scikit-learn==1.3.0
scipy==1.11.3
matplotlib==3.7.2

# ========================================
# UTILITIES
# ========================================
PyJWT==2.8.0
wheel==0.41.2

# ========================================
# DEVELOPMENT TOOLS (Optional)
# ========================================
# pytest==7.4.2
# black==23.7.0
# flake8==6.0.0

# ========================================
# INSTALLATION NOTES
# ========================================
# 1. Use Python 3.10.1 ONLY (not 3.11+ due to MediaPipe conflicts)
# 2. Create fresh virtual environment: python -m venv venv_py310
# 3. Activate: venv_py310\Scripts\activate (Windows)
# 4. Upgrade pip first: python -m pip install --upgrade pip==23.2.1  
# 5. Install: pip install -r requirements_exact.txt
# 6. Verify: python -c "import tensorflow, mediapipe, cv2; print('✅ All imports successful')"
#
# CONFLICT PREVENTION:
# - Never use "pip install package" without version numbers
# - Never upgrade individual packages - reinstall entire requirements
# - If errors occur, delete venv and recreate from scratch
# - Windows users: Install Visual C++ Redistributable 2019+